#6 feat: Enhance process signal handling and database connection management
- Added handling for SIGINT and SIGTERM signals to gracefully flush buffered logs and close all database connections before exiting.
- Moved SIGINT handling from start.ts to index.ts for centralized signal management.
- Integrated DatabaseService to ensure all connections are properly closed during unexpected exits.
